add windows support
This patch adds support for popping open Visual Studio
to debug on windows.

The implementation is factored out into an internal module
to make it clearer that not all `pub` functions are part of
the actual public api of the module.

The docs are updated to reflect the new reality.

I also realised that std::process::id is a thing while
doing this port, so I was able to drop the dependency on
`nix` entirely.

Closes #7